static_cast

+++ Core/AppRuntime/Source/AppRuntimeWin32.cpp
@@ -20,7 +20,7 @@ namespace Babylon
         auto coInitScopeGuard = gsl::finally([] { CoUninitialize(); });
 
         char filename[FILENAME_BUFFER_SIZE];
-        auto result = GetModuleFileNameA(nullptr, filename, std::size(filename));
+        auto result = GetModuleFileNameA(nullptr, filename, static_cast<DWORD>(std::size(filename)));
         assert(result != 0);
         (void)result;
         RunEnvironmentTier(filename);
